+++ src/Git/Plantation/Cmd/Member.hs
@@ -1,12 +1,21 @@
 {-# LANGUAGE DataKinds        #-}
+{-# LANGUAGE LambdaCase       #-}
 {-# LANGUAGE OverloadedLabels #-}
 {-# LANGUAGE TypeOperators    #-}
 
-module Git.Plantation.Cmd.Member where
+module Git.Plantation.Cmd.Member
+  ( MemberCmdArg
+  , MemberArg
+  , actForMember
+  , inviteUserToRepo
+  , kickUserFromRepo
+  ) where
 
 import           RIO
 
+import           Data.Aeson.Text                      (encodeToLazyText)
 import           Data.Extensible
+import           Git.Plantation.Cmd.Arg
 import           Git.Plantation.Cmd.Repo              (repoGithub,
                                                        splitRepoName)
 import           Git.Plantation.Data
@@ -15,57 +24,74 @@ import           GitHub.Data.Name                     (mkName)
 import qualified GitHub.Endpoints.Repos.Collaborators as GitHub
 import qualified Mix.Plugin.GitHub                    as MixGitHub
 
-type InviteMemberCmd = Record MemeberCmdFields
-
-type KickMemberCmd = Record MemeberCmdFields
-
-type MemeberCmdFields =
-  '[ "team"  >: Text
-   , "repos" >: [Int]
-   , "user"  >: Maybe Text
+type MemberCmdArg = Record
+  '[ "team"  >: TeamId
+   , "repos" >: [RepoId]
+   , "user"  >: Maybe UserId
    ]
 
-actForMember :: (User -> Repo -> Plant ()) -> Record MemeberCmdFields -> Team -> Plant ()
-actForMember act args team = case args ^. #repos of
-  [] -> forM_ (team ^. #repos) $ \repo -> forM_ member $ act' repo
-  _  -> forM_ repos            $ \repo -> forM_ member $ act' repo
-  where
-    act' repo user = tryAnyWithLogError $ act user repo
-    repos  = catMaybes $ flip lookupRepoByProblemId team <$> args ^. #repos
-    member = maybe (team ^. #member) (: []) $ flip lookupUser team =<< args ^. #user
+type MemberArg = Record
+  '[ "user" >: User
+   , "repo" >: Repo
+   ]
 
-inviteUserToRepo :: User -> Repo -> Plant ()
-inviteUserToRepo user target = do
-  github <- repoGithub target
+actForMember :: (MemberArg -> Plant ()) -> MemberCmdArg -> Plant ()
+actForMember act args =
+  findByIdWith (view #teams) (args ^. #team) >>= \case
+    Nothing   -> logError $ display (encodeToLazyText $ errMsg $ args ^. #team)
+    Just team -> do
+      member <- findMember (args ^. #user) team
+      repos  <- findRepos (args ^. #repos) team
+      mapM_ act $ hsequence $ #user <@=> member <: #repo <@=> repos <: nil
+
+findMember :: Maybe UserId -> Team -> Plant [User]
+findMember Nothing team = pure $ team ^. #member
+findMember (Just idx) team =
+  case findById idx (team ^. #member) of
+    Nothing   -> logError (display $ encodeToLazyText $ errMsg idx) >> pure []
+    Just user -> pure [user]
+
+findRepos :: [RepoId] -> Team -> Plant [Repo]
+findRepos [] team = pure $ team ^. #repos
+findRepos ids team = fmap catMaybes . forM ids $ \idx ->
+  case findById idx (team ^. #repos) of
+    Nothing -> logError (display $ encodeToLazyText $ errMsg idx) >> pure Nothing
+    Just r  -> pure (Just r)
+
+inviteUserToRepo :: MemberArg -> Plant ()
+inviteUserToRepo args = do
+  github <- repoGithub $ args ^. #repo
   let (owner, repo) = splitRepoName github
   resp <- MixGitHub.fetch $ \auth -> GitHub.addCollaborator auth
     (mkName Proxy owner)
     (mkName Proxy repo)
-    (mkName Proxy $ user ^. #github)
+    (mkName Proxy $ args ^. #user ^. #github)
   case resp of
-    Left err -> logDebug (displayShow err) >> throwIO (InviteUserError err user target)
+    Left err -> logDebug (displayShow err) >> throwIO (failure err)
     Right _  -> logInfo $ display (success github)
   where
+    failure err = InviteUserError err (args ^. #user) (args ^. #repo)
     success githubPath = mconcat
       [ "Success: invite "
-      , user ^. #name, "(", user ^. #github, ")"
+      , args ^. #user ^. #name, "(", args ^. #user ^. #github, ")"
       , " to ", githubPath, "."
       ]
 
-kickUserFromRepo :: User -> Repo -> Plant ()
-kickUserFromRepo user target = do
-  github <- repoGithub target
+kickUserFromRepo :: MemberArg -> Plant ()
+kickUserFromRepo args = do
+  github <- repoGithub $ args ^. #repo
   let (owner, repo) = splitRepoName github
   resp <- MixGitHub.fetch $ \auth -> GitHub.removeCollaborator auth
     (mkName Proxy owner)
     (mkName Proxy repo)
-    (mkName Proxy $ user ^. #github)
+    (mkName Proxy $ args ^. #user ^. #github)
   case resp of
-    Left err -> logDebug (displayShow err) >> throwIO (KickUserError err user target)
+    Left err -> logDebug (displayShow err) >> throwIO (failure err)
     Right _  -> logInfo $ display (success github)
   where
+    failure err = KickUserError err (args ^. #user) (args ^. #repo)
     success githubPath = mconcat
       [ "Success: kick "
-      , user ^. #name, "(", user ^. #github, ")"
+      , args ^. #user ^. #name, "(", args ^. #user ^. #github, ")"
       , " from ", githubPath, "."
       ]